\name{uploadTwittR-package}
\alias{uploadTwittR-package}
\alias{uploadTwittR}
\docType{package}
\title{
\packageTitle{uploadTwittR}
}
\description{
\packageDescription{uploadTwittR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{uploadTwittR}
\packageIndices{uploadTwittR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{uploadTwittR}

Maintainer: \packageMaintainer{uploadTwittR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
